Context now panics if it requests too much memory
Before Context.growMem would attempt to allocate all memory of the users system if memory size increases beyond that which is visible by a 32-bit system. This is because the internals is transpiled from 32-bit wasm binary.

Now specifying a maximum heap size limit will cause the context to panic instead of blindly allocating more memory.